==> 3_RAS84_patient_classification_TCGA_LUAD.Rmd <==
---
title: "RAS84 signature sample classification of TCGA LUAD cohort : Project 002"
author: "philip.east@crick.ac.uk"
date: "12/02/2019"
output: 
  html_document:
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script classifies TCGA samples using the RAS84 signature
constructed during the validate_signatures_CCLE
analysis. Classification groups are described and collorates across
the groupings are reported.

```{r r_init,message=FALSE}
library( affy )
library( limma )
library( reshape2 )
library( scales )
library( RColorBrewer )
library( ComplexHeatmap )
library( gplots )
library( gridExtra )
library( ggthemes )
library( Hmisc )
library( vcd )
library( UpSetR )
library( monocle )
library( pls )
library( DESeq2 )
library( circlize )
library( ggforce )
library( openxlsx )
library( tidyverse )
library( ggrepel )

my_grey <- "#707173"
my_red <- "#993333"
my_orange <- "#f6ad6e"
my_green <- "#7ab51d"
my_lightgreen <- "#adcf82"
my_purple <- "#bb90bd"
my_blue <- "#336699"

plot_formatter <- function() {
    theme_bw( ) +
        theme( panel.grid.major = element_blank( ),
               panel.grid.minor = element_blank( ),
               panel.border = element_blank( ),
              panel.background = element_blank( ),
              axis.line = element_line(color = "black"),
              axis.line.x = element_line(color="black", size = 0.5 ),
              axis.line.y = element_line(color="black", size = 0.5 ),
              text = element_text( size = 12 ) )
}

tumour_type <- "LUAD"

## data resources
TCGA_LUAD_SE_FILE <- file.path( "data", "objects", paste0( "se_t_", tumour_type, ".RNA-Seq.legacy.biolinks.rda" ) )
RAS_SIG_HOME <- file.path( "data", "resources", "signatures", "ras" )
RAS84_FEATURE_MAP_FILE <- file.path( "data", "resources", "RAS84_feature_map.xlsx" )
TCGA_LUAD_RAS84_CLASS_SE_FILE <- file.path( "data", "objects", "TCGA_LUAD_RAS84_class_se_t.rds" )

## figures
TCGA_LUAD_RAS84_SIGNATURE_CLASS_OVERLAP_FILE <- file.path( "figures", "TCGA_LUAD_RAS84_signature_class_overlap_TCGA.pdf" )

```

# Introduction

Here we classify TCGA `r tumour` patients using our RAS84 expression
signature. We use a hierarchical clustering approach with a ward.D2
agglomeration method. We find that five clusters produced a low RAS
activity cluster with a very low KRAS mutation rate. We go on to show
other oncogenic mutations that correlate with this clustering pattern.

# Project Settings

We run the classification using the RAS84 signature against the LUAD
cohort from TCGA. We produce five RAS activity patient clusters.

# Load and process TCGA data

Load the TCGA LUAD SummarizedExperiment object from the GDC
legacy download. These objects contain the TCGA data as downloaded. We
go on to add sample classification labels. 

```{r tcga_se}
se_t <- readRDS( TCGA_LUAD_SE_FILE )
```

# RAS Signatures

Read in the RAS84 signature genes and the parent signatures.

```{r ras84_sig}
krasSigNames <- sub( "\\.csv", "", dir( RAS_SIG_HOME ) )
signatures <- lapply( krasSigNames, function( sigFile ) {
    read.table( file = file.path( RAS_SIG_HOME, paste( sigFile, "csv", sep = "." ) ),
               sep = "\t", stringsAsFactors = FALSE )[, 1 ]
} )
names( signatures ) <- make.names( krasSigNames )
ras84_map <- read.xlsx( RAS84_FEATURE_MAP_FILE )
signatures$RAS84 <- ras84_map$TCGA_gene_symbol
```

# Classify samples

We classify the samples using hierarchical clustering approach with a
ward.D2 agglomeration method. We use cutree to cut the resulting
dendrogram into five clusters. We label the clusters RAG0-4 depending
on the mean expression across the clusters, with RAG4 being the highest.

```{r classification}
samp_clust_l <- lapply( signatures, function( sig_genes ) {
    sig_f <- rowData( se_t )$gene_symbol %in% sig_genes
    vst_mat <- assays( se_t )$vst[ sig_f, ]
    hc <- hclust( dist( t( vst_mat ) ), method = "ward.D2" )
    ct_df <- cutree( hc, k = cluster_no ) %>%
        as.data.frame( ) %>%
        dplyr::rename( ct_cluster = 1 ) %>%
        rownames_to_column( var = "barcode" )
    ct_means <- vst_mat %>%
        as.data.frame( ) %>%
        tidyr::gather( barcode, value ) %>%
        left_join( ct_df, by = "barcode" ) %>%
        group_by( ct_cluster ) %>%
        dplyr::summarize( mean = mean( value ) ) %>%
        arrange( mean ) %>%
        mutate( RAG_5 = paste( "RAG", 0:( cluster_no - 1 ), sep = "-" ) )
    ct_df %>%
        left_join( ct_means, by = "ct_cluster" ) %>%
        mutate( RAG_5 = factor( RAG_5 ) ) %>%
        dplyr::select( barcode, RAG_5 )
} )

samp_clust_df <- map( samp_clust_l, function( x ) x$RAG_5 ) %>%
    bind_cols( ) %>%
    as.data.frame( )
samp_clust_df$barcode <- samp_clust_l[[ 1 ]]$barcode

colData( se_t ) <- colData( se_t ) %>%
    as.data.frame( ) %>%
    left_join( samp_clust_df, by = "barcode" ) %>%
    DataFrame( row.names = .$barcode )
```
```{r class_sig_overlap_heatmap}
hm_dat <- samp_clust_df %>%
    dplyr::select( -barcode ) %>%
    relocate( RAS84, .before = 1 ) %>%
    relocate( Loboda, .before = 3 )

for( i in colnames( hm_dat ) ) {
    hm_dat[, i ] <- as.numeric( hm_dat[, i ] )
}
    
hm <- Heatmap( hm_dat,
              name = "Overlap", 
              clustering_method_rows = "ward.D2",
              show_column_names = TRUE,
              col = c( "grey" , "darkgreen", "purple", "orange", "black" ),
              row_split = hm_dat$RAS84,
              column_split = factor( c( "RAS84", rep( "public_sig", 4 ) ), levels = c( "RAS84", "public_sig" ) ),
              column_title = sig_n,
              column_title_gp = gpar( fontsize = 14 ),
              column_names_gp = gpar( fontsize = 12 ),
              heatmap_legend_param = list( title_gp = gpar( fontsize = 14 ),
                                          labels_gp = gpar( fontsize = 12 ) ),
              show_heatmap_legend = TRUE,
              cluster_row_slices = FALSE,
              row_names_max_width = unit( 10, "mm" ) )

cairo_pdf( file = TCGA_LUAD_RAS84_SIGNATURE_CLASS_OVERLAP_FILE, width = 4 )
print( hm )
dev.off( )
```

# Heatmaps

Heatmap showing the expression of the RAS84 signature across the five clusters. Patient KRAS mutation status is shown for each sample row indicated in purple. Parent signature gene mappings are highlighed at the bottom in grey.

## Sample mutation annotation

```{r mut_anno}
anno_row_df <- colData( se_t ) %>%
    as.data.frame( ) %>%
    mutate( kras_mut = onco_ras == "kras_mut" ) %>%
    dplyr::select( kras_mut )
cols_row <- list( kras_mut = c( 'TRUE' = my_purple, 'FALSE' = "white" ) )
hm_anno_row <- rowAnnotation( df = anno_row_df,
                             col = cols_row,
                             na_col = "white",
                             gap = unit( 1, "mm" ),
                             show_legend = FALSE,
                             show_annotation_name = TRUE,
                             annotation_name_side = "top",
                             annotation_legend_param = list( title_gp = gpar( fontsize = 12 ),
                                                            labels_gp = gpar( fontsize = 10 ) ),
                             simple_anno_size = unit( 3, "mm" ) )
```

## Heatmap

```{r render_heatmap}
hmcols <- colorpanel( 1000, my_blue, "white", my_red )
sig_n <- "RAS84"
sig_f <- rowData( se_t )$gene_symbol %in% signatures[[ sig_n ]]
hmdat <- assays( se_t )$vst[ sig_f, ] %>%
                      as.data.frame( ) %>%
                      rownames_to_column( var = "id" ) %>%
                      left_join( as.data.frame( rowData( se_t )[, c( "id", "gene_symbol" ) ] ), by = "id" ) %>%
                      dplyr::select( -id ) %>%
                      gather( barcode, value, -gene_symbol ) %>%
                      spread( gene_symbol, value ) %>%
                      left_join( samp_clust_l[[ sig_n ]], by = "barcode" ) %>%
                      column_to_rownames( var = "barcode" )
hmdat <- hmdat[ rownames( anno_row_df ), ]
    hmmat <- hmdat %>%
        dplyr::select( -RAG_5 ) %>%
        as.matrix( ) %>%
        t( )
hmmat <- hmmat - rowMedians( hmmat )
hm <- Heatmap( t( hmmat ),
              name = "Expression", 
              clustering_method_rows = "ward.D2",
              clustering_method_columns = "ward.D2",
              show_column_names = TRUE,
              show_row_names = FALSE,
              col = colorRamp2(
                  seq( from = -2, to = 2, length.out = length( hmcols ) ),
                  hmcols ),
              row_split = hmdat$RAG_5,
              column_title = sig_n,
              column_title_gp = gpar( fontsize = 14 ),
              column_names_gp = gpar( fontsize = 4 ),
              heatmap_legend_param = list( title_gp = gpar( fontsize = 14 ),
                                          labels_gp = gpar( fontsize = 4 ) ),
              show_heatmap_legend = FALSE,
              cluster_row_slices = FALSE,
              row_names_max_width = unit( 10, "mm" ) )
print( hm + hm_anno_row )
```

# RAS Index

We calculate a RAS index value or RAS84 signature mean value for each
sample. These values are used to represent overall RAS signalling
activivty and are used in linear regression analysis and to rank
samples based on RAS activity.

```{r ras_index}
ras_index <- assays( se_t )$vst %>%
                          as.data.frame( ) %>%
                          rownames_to_column( var = "id") %>%
                          tidyr::gather( barcode, value, -id ) %>%
                          left_join( as.data.frame( rowData( se_t ) ),
                                    by = "id" ) %>%
                          filter( RAS84 ) %>%
                          group_by( barcode ) %>%
                          dplyr::summarize( ras_index = mean( value ) )
  
colData( se_t ) <- colData( se_t ) %>%
    as.data.frame( ) %>%
    left_join( ras_index, by = "barcode" ) %>%
    DataFrame( row.names = .$rowname )

gg <- colData( se_t ) %>%
  as.data.frame( ) %>%
    ggplot( aes_string( x = "RAS84", y = "ras_index" ) ) +
  geom_violin( ) +
  geom_boxplot( width = 0.2 ) +
  labs( y = "RAS84 expression", x = "RAS activity" )
gg <- my_gg_themes( gg )
gg <- gg + theme( axis.text.x = element_text( angle = 90, hjust = 0.95 ) ) +
  theme( text = element_text( size = 14 ) )
gg

saveRDS( se_t, file = TCGA_LUAD_RAS84_CLASS_SE_FILE )
```

# TNM Stage across RAS groups

Here we look at tumour stage across our RAS groups.

```{r stage,echo=FALSE,results='hide'}
substage_to_stage <- data.frame( tumor_stage = c( 'stage i', 'stage ia', 'stage ib', 'stage ii',
                                              'stage iia', 'stage iib', 'stage iiia' , 'stage iiib',
                                              'stage iv' ),
                                 stage = c( 'I', 'I', 'I', 'II', 'II', 'II', 'III', 'III', 'IV' ) )
                       
plotdat <- colData( se_t ) %>%
    as.data.frame( ) %>%
    select( RAS84, tumor_stage ) %>%
    left_join( substage_to_stage, by = "tumor_stage" )

stage_gg <- ggplot( plotdat, aes( x = RAS84, fill = stage ) ) +
    geom_bar( position = "fill" ) +
    scale_y_continuous( labels = percent_format( ) ) +
    theme( axis.text.x = element_text( angle = 90, vjust = 0.5, hjust=1 ) )

print( stage_gg )
dev.off( )
substage_gg <- ggplot( plotdat, aes( x = RAS84, fill = tumor_stage ) ) +
    geom_bar( )
print( substage_gg )

stage_tab <- table( plotdat[, c( "RAS84", "stage" ) ] )
chisq.test( stage_tab )

mosaic( stage_tab, shade = TRUE, legend = TRUE, main = "tumour stage",
       gp = shading_hsv,
       gp_args = list( h = c(28/360, 203/360 ), s = c( 1, 0 ), v = c( 1, 0.5 ),
                      interpolate = c( 2, 4 ) ) )
```

# KRAS mutation distributions

We assess each signature by its ability to discriminate KRAS mutation
between the high and low groups. Here we use a chisq-test to assign
significance to the distribution of KRAS mutants across the high and
low groups. Patients with oncogenic RAS-pathway mutations, other than KRAS are excluded as in the CCLE analysis. 

```{r kras_dist_pval}
tab_input_dat <- map( samp_clust_l, left_join,
                     as.data.frame( colData( se_t )[, c( "barcode", "MUT.KRAS" ) ] ),
                     by = "barcode" ) %>%
    map( select, -barcode ) %>%
    map( table ) %>%
    map( chisq.test ) %>%
    map( function( x ) x$p.value )
```

```{r kras_dist_pval_plot,fig.width=3,fig.height=3,results='hide',warning=FALSE}
rgm <- assays( se_t )$vst %>%
                    as.data.frame( ) %>%
                    rownames_to_column( var = "id" ) %>%
                    filter( id %in% RAS84_TCGA_ids ) %>%
                    tidyr::gather( barcode, value, -id )
new_pvalfig_gg <- melt( tab_input_dat ) %>%
    rename( signatures = L1 ) %>%
    mutate( group = factor( c( "selected", "selected",
                              "discarded", "selected", "control", "discarded",
                              "merged" ),
                           levels = c( "merged", "selected", "discarded", "control" ) ) ) %>%
    ggplot( aes( x = 1, y = -log10( value ),
               color = group, label = signatures ) ) +
    geom_vline( xintercept = 1 ) +
    geom_point( size = 5 ) +
    geom_text_repel( nudge_x = 0.1 ) +
    plot_formatter( ) +
    scale_color_manual( values = c( my_purple, my_blue, my_green, my_grey ) ) +
    theme( legend.position = "none",
          text = element_text( size = 8 ) ) 
print( new_pvalfig_gg )
```
