==> 7_RAS84_pancancer_survival_analysis.Rmd <==
---
title: "Pancancer : Project 002"
author: "philip.east@crick.ac.uk"
date: 'Compiled: `r format(Sys.time(), "%d %B, %Y @ %H:%M:%S")`'
output:
  html_document:
    df_print: tibble33
    toc: true
    toc_depth: 5
    toc_float: true
    code_folding: hide
    number_sections: yes
---

```{css setup_css, echo=FALSE}
body .main-container {
  max-width: 80%;
}
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Introduction

Presented here is a pan-TCGA RAS activity analysis. First we calculate
a RAS index value for each tumour (RAS84 signature mean) across a
pan-TCGA normalised expression data set. Second we classify
tumours, from each TCGA tumour cohort, into three RAS activity groups
(RASsig_max, RASsig_1, and RASsig_0). The RAS index distributions for
each tumour and each RAS activity group are shown below. We then
explore the link between RAS activity and overall survival for each
tumour type. The hypothesis being that RAS activty, in the more RAS
driven tumours (where the signature is likely to be more effective), is more
likely to correlate with overall survival. 

```{r r_init,message=FALSE,warning=FALSE}
library( affy )
library( limma )
library( org.Hs.eg.db )
library( ggplot2 )
library( reshape2 )
library( scales )
library( RColorBrewer )
library( ComplexHeatmap )
library( circlize )
library( colorspace )
library( GetoptLong )
library( dendextend )
library( matrixStats )
library( DESeq2 )
library( gplots )
library( ggplot2 )
library( gridExtra )
library( tibble )
library( dplyr )
library( tidyr )
library( gtable )
library( openxlsx )
library( survival )
library( survminer )
library( ggrepel )
library( DT )
library( broom )
library( purrr )


my_grey <- "#707173"
my_red <- "#e3001a"
my_orange <- "#f6ad6e"
my_green <- "#7ab51d"
my_lightgreen <- "#adcf82"
my_purple <- "#bb90bd"
my_blue <- "#4066aa"

plot_formatter <- function() {
    theme_bw( ) +
        theme( panel.grid.major = element_blank( ),
               panel.grid.minor = element_blank( ),
               panel.border = element_blank( ),
               panel.background = element_blank( ),
               text = element_text( size = 10 ) )
}

## data resources
TCGA_PANCANCER_SE_FILE <- file.path( "data", "objects", "pancancer_se.rds" )

## figures
KMPLOT_PANCANCER_TUMOUR_PLOT_FILE <- file.path( "figures", "kmplot_tumour.pdf" )
KMPLOT_RI_PANCANCER_TUMOUR_PLOT_FILE <- file.path( "figures", "kmplot_RI_tumour.pdf" )
```

# Load tumour SummarizedExperiment objects

```{r se_objects}
se_pc <- readRDS( TCGA_PANCANCER_SE_FILE )
```

# Survival data

Some of the tumours have low survival events. The distributions for OS
and PFI are shown below. 

## Number of events per tumour {.tabset .tabset-fade}

```{r cell_imm_lscape,results='asis',echo=FALSE,message=FALSE,warning=FALSE}
for( i in c( "OS", "PFI" ) ) {
    cat( "###", i, "\n" )
    event_dat <- colData( se_pc ) %>%
        as.data.frame( ) %>%
        mutate_( event = i ) %>%
        filter( !is.na( event ) ) %>%
        group_by( tumour ) %>%
        summarize( events = sum( event == 1 ) )

    gg <- ggplot( event_dat, aes( x = tumour, y = events ) ) +
        geom_bar( stat = "identity" ) +
        theme( axis.text.x = element_text( angle = 90, hjust = 1 ) )
    gg
    cat( "\n\n" )
}
```

## Filter low event tumours

The TCGA-CDR paper advises not to use OS with DLBC, PCPG, TGCT and
THYM, and caution with BRCA, KICH, LGG, PRAD, READ, and THCA.

For PFI, they advise not to use it with LAML, and PCPG, and caution
with DLBC and KICH.

We removed these tumour types prior to the relevent survival data type
analysis.

We also only select tumours in our high RAS group.

```{r tumour__filter}
high_ras_tumour_group <- c( "TCGA-PAAD", "TCGA-COAD", "TCGA-READ", "TCGA-LUAD",
                           "TCGA-STAD", "TCGA-BLCA", "TCGA-HNSC", "TCGA-CESC",
                           "TCGA-LUSC", "TCGA-ESCA" )
surv_tumour_exclude <- list( OS = c( "TCGA-DLBC", "TCGA-PCPG", "TCGA-TGCT", "TCGA-BRCA", "TCGA-KICH",
                                    "TCGA-LGG", "TCGA-PRAD", "TCGA-READ", "TCGA-THCA" ),
                            PFI = c( "TCGA-LAML", "TCGA-PCPG", "TCGA-DLBC", "TCGA-KICH" ) )
surv_endpoint <- "OS"
endpoint_time <- paste0( surv_endpoint, ".Time" )
se_l_endpoint <- se_l[ !names( se_l ) %in% surv_tumour_exclude[[ surv_endpoint ]] ]
n_count <- unlist( lapply( se_l_endpoint, ncol ) )
se_l_endpoint <- se_l_endpoint[ n_count > 100 ]
se_l_endpoint <- se_l_endpoint[ names( se_l_endpoint ) %in% high_ras_tumour_group ]
```

# Survival analysis

There are two options open to us here. We can determine if there is
any correlation with overall survival across our RAS activity
groups. This is attractive since it is easy to construct KM plots when
the data are grouped. It does rely on the accurate classification of
the samples in terms of RAS activity. This may well not be the case as
we have demonstrated in LUAD where three clusters was not
sufficient to capture all of the complexity in the expression of the
RAS signature. The other option is to run a cox regression analysis
against the RAS index values. This does not rely on the grouping of the data
but may fail if there is a non-linear relationship between RA index
and outcome. Presented below are both approaches.

## Quantile survival analysis

Here we split the samples based on quantiles of the principle components. We then run survival analysis between the top and bottom quantile.

```{r quantile_survival}
cox_reg_l <- lapply( se_l_endpoint, function( se ) {
    tumour <- se$tumour[ 1 ]
    cph <- coxph( OS.Time ~ OS ~ ras_index, as.data.frame( colData( se ) ) )
    as.data.frame( summary( cph )$coef )
} )
```

## RAS index regression results

Here we only look at the coxph regression results for RAS index

```{r ras_index_res}
decide_ri <- bind_rows( cox_reg_l, .id = "tumour" ) %>%
    mutate( padj = p.adjust( pval, method = "BH" ) ) %>%
    arrange( padj )

form <- as.formula( paste0( "Surv( ", endpoint_time, "/365*12, ", surv_endpoint, " ) ~ ras_index" ) )

for( i in decide_ri$tumour ) {
    se <- se_l_endpoint[[ i ]]
    fit <- coxph( form, data = colData( se ) )
    surv_l <- list( mid = survfit( fit, newdata = data.frame( ras_index = mean( se$ras_index) ),
                                  data = as.data.frame( colData( se ) ) ),
                   upper = survfit( fit, newdata = data.frame( ras_index = mean( se$ras_index ) + 1 ),
                                   data = as.data.frame( colData( se ) ) ),
                   lower = survfit( fit, newdata = data.frame( ras_index = mean( se$ras_index ) - 1 ),
                                   data = as.data.frame( colData( se ) ) ) )
    gg <- ggsurvplot( surv_l,
               combine = TRUE,
               palette = c( my_grey, my_red, my_blue ) )
    cairo_pdf( file = sub( "tumour", i, KMPLOT_RI_PANCANCER_TUMOUR_PLOT_FILE ),
              width = 4, height = 4 )
    print( gg )
    dev.off( )
}
```
